+++ packages/hook/src/config.d.json.ts
@@ -1,14 +1,29 @@
-declare const config: {
-	exposeEndpoint: boolean
-	endpointPath: string
-	secureEndpoint: boolean
+declare const config: (
+	| {
+			exposeEndpoint: true
+			endpointPath: string
+			secureEndpoint: boolean
+	  }
+	| {
+			exposeEndpoint: false
+			endpointPath?: never
+			secureEndpoint?: never
+	  }
+) & {
 	tsSchema: BaseSchemaConfig & { includeDocs: boolean }
 	zodSchema: BaseSchemaConfig
 }
 
-type BaseSchemaConfig = {
-	generateFile: boolean
-	outputPath: string
+type BaseSchemaConfig = (
+	| {
+			generateFile: true
+			outputPath: string
+	  }
+	| {
+			generateFile: false
+			outputPath?: string
+	  }
+) & {
 	includeSystemCollections: boolean
 	banner?: string[]
 	overrides?: {