+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+
+## 0.7.0
+
 - remove `importStatements` option in favour of `banner`
 - make options like `outputPath` optional when they're not needed
 - make zod schema for `collectionId` and `collectionName` stricter (fixes #22)